Add a title_exists() method to the Entry class
Testing the PdfFileReader().getDocumentInfo() method revealed an important
discovery. Sometimes, the method will return a dictionary without the
['/Title'] key. This probably happens in the event a PDF lacks title metadata.

The new Entry.title_exists() method checks the validity of
Entry.metadata['/Title']. The method returns false either if the key contains
an empty string or if the key does not exist altogether.